+++ src/index.css
@@ -1,3 +1,15 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+@layer base {
+  body {
+    margin: 0;
+    padding: 0;
+    height: 100vh;
+  }
+
+  #root {
+    height: 100%;
+  }
+}
